refactor: update package structure and replace models with DTOs for improved separation of concerns
