+++ brawl-core/src/main/java/brawl/example/project_brawl_api_sheets/integration_sheets/brawl_sheets/dto/TeamBattleResponseDTO.java
@@ -1,6 +1,5 @@
 package brawl.example.project_brawl_api_sheets.integration_sheets.brawl_sheets.dto;
 
-import brawl.example.project_brawl_api_sheets.integration_sheets.brawl_sheets.model.BattleLog;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -8,7 +7,7 @@ import java.util.*;
 
 @Data
 @AllArgsConstructor
-public class TeamBattleDTO {
+public class TeamBattleResponseDTO {
     private String teamName;
-    private List<BattleLog.BattleLogInfo> battles;
+    private List<BattleLogReceiveDTO.BattleLogInfo> battles;
 }
